Moved and split out the nav bar
